requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
ecdsa==0.19.0
fastapi==0.115.8
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
redis==5.0.1
python-jose==3.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
pytest-cov==4.1.0
AsyncMock==1.0.0
